NETOBSERV-2760: [upstream] manage k8s client calls for status
Backend:
- Make ConsoleMode an explicit config (Sandalone vs OpenShiftPlugin vs
  Mock)
- K8S config now takes token mode, just like prom & loki (it was
  previously forcing token forwarding, which currently cannot work in
standalone mode)
- new K8S client call to get FlowCollector status (used only in
  standalone mode; as a console plugin, it continues to use the console
build-in helpers like)
- little refactoring of the token extraction logic, now re-used between
  loki client and k8s client.

Frontend:
- Renamed moduleMapper/dummy.tsx => pluginToStandaloneMapper/mapper.tsx
- In this mapper, now "useK8sWatchResource" only mocks when tthe console
  is configured in Mock mode; when in Standalone mode, it peforms an
actual k8s query.
- Status indicator now shows the actual error rather than a static
  message.
