+++ README.md
@@ -195,4 +195,4 @@ When you update the operator CRDs, you may have to update some schemas here as w
 
 When a CRD field is added, consider whether you need to update these files.
 
-Additionally, [schema.ts](./web/moduleMapper/schemas.ts) contains the full CRD schema as JSON, used in tests and in standalone mode, and should also be kept up to date. Use [generate-schemas.sh](./scripts/generate-schemas.sh) to regenerate them (you need a running cluster with the CRDs installed).
+Additionally, [schema.ts](./web/pluginToStandaloneMapper/schemas.ts) contains the full CRD schema as JSON, used in tests and in standalone mode, and should also be kept up to date. Use [generate-schemas.sh](./scripts/generate-schemas.sh) to regenerate them (you need a running cluster with the CRDs installed).
