+++ pkg/kubernetes/resources/resources.go
@@ -10,6 +10,23 @@ import (
 	"k8s.io/client-go/rest"
 )
 
+func Get(ctx context.Context, token, namespace, name string, gvr schema.GroupVersionResource) (*unstructured.Unstructured, error) {
+	config, err := rest.InClusterConfig()
+	if err != nil {
+		return nil, err
+	}
+	config.BearerToken = token
+	config.BearerTokenFile = ""
+
+	dynamicClient, err := dynamic.NewForConfig(config)
+	if err != nil {
+		return nil, err
+	}
+
+	// Retrieve the custom resource
+	return dynamicClient.Resource(gvr).Namespace(namespace).Get(ctx, name, v1.GetOptions{})
+}
+
 func List(ctx context.Context, token string, gvr schema.GroupVersionResource) ([]unstructured.Unstructured, error) {
 	config, err := rest.InClusterConfig()
 	if err != nil {
