+++ web/pluginToStandaloneMapper/templates.ts
