+++ web/src/model/config.ts
@@ -37,7 +37,7 @@ export type Config = {
   features: Feature[];
   fields: FieldConfig[];
   dataSources: string[];
-  lokiMocks: boolean;
+  consoleMode: 'Standalone' | 'OpenShiftPlugin' | 'Mock';
   lokiLabels: string[];
   promLabels: string[];
   maxChunkAgeMs?: number;
@@ -62,7 +62,7 @@ export const defaultConfig: Config = {
   features: [],
   fields: [],
   dataSources: ['loki', 'prom'],
-  lokiMocks: false,
+  consoleMode: 'Standalone',
   lokiLabels: [],
   promLabels: [],
   maxChunkAgeMs: undefined,
